requirements.txt
Flask==3.0.3 
Werkzeug==3.0.3
bcrypt==4.1.3 
google-cloud-firestore==2.16.0 
google-cloud-storage==2.16.0
ultralytics==8.2.18 
opencv-python-headless==4.9.0.80 
torch==2.3.0 
torchvision==0.18.0 
Pillow==10.3.0
python-dotenv==1.0.1